+++ src/main/java/com/app/smartdrive/api/repositories/master/CityRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CityRepository extends JpaRepository<Cities, Integer> {
+public interface CityRepository extends JpaRepository<Cities, Long> {
 }
