+++ src/main/java/com/app/smartdrive/api/repositories/master/IbmeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IbmeRepository extends JpaRepository<InboxMessaging, Integer> {
+public interface IbmeRepository extends JpaRepository<InboxMessaging, Long> {
 }
